refactor(context-menu): H.4 — extract MultiAlbumToPlaylistSubmenu
